+++ changelog.md
@@ -1,2 +1,2 @@
-# Fabric
-- fix NullPointerException when trying to place a sponge in water
+# Paper
+- add option to require a player to have a permission ('magnetic.ability.use') to use the ability